+++ src/main/java/com/darksci/pardot/api/response/visitor/VisitorQueryResponse.java
@@ -15,22 +15,52 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.darksci.pardot.api.parser.visitor;
+package com.darksci.pardot.api.response.visitor;
 
-import com.darksci.pardot.api.parser.JacksonFactory;
-import com.darksci.pardot.api.parser.ResponseParser;
-import com.darksci.pardot.api.response.visitor.Visitor;
-import com.darksci.pardot.api.response.visitor.VisitorReadResponse;
+import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
 
-import java.io.IOException;
+import java.util.List;
 
 /**
- * Handles parsing Visitor Read API responses into POJOs.
+ * Represents the result from a Visitor Query API call.
  */
-public class VisitorReadResponseParser implements ResponseParser<Visitor> {
+public class VisitorQueryResponse {
+    private VisitorQueryResponse.Result result;
+
+    public Result getResult() {
+        return result;
+    }
+
+    /**
+     * Represents one or more visitors.
+     */
+    public static class Result {
+        private Integer totalResults = 0;
+
+        @JacksonXmlProperty(localName = "visitor")
+        private List<Visitor> visitors;
+
+        public Integer getTotalResults() {
+            return totalResults;
+        }
+
+        public List<Visitor> getVisitors() {
+            return visitors;
+        }
+
+        @Override
+        public String toString() {
+            return "Result{"
+                + "totalResults=" + totalResults
+                + ", visitors=" + visitors
+                + '}';
+        }
+    }
 
     @Override
-    public Visitor parseResponse(final String responseStr) throws IOException {
-        return JacksonFactory.newInstance().readValue(responseStr, VisitorReadResponse.class).getVisitor();
+    public String toString() {
+        return "VisitorQueryResponse{"
+            + "result=" + result
+            + '}';
     }
 }